refactor: TitleAndUUID/DocumentTitle 타입을 공통 타입 파일로 이동
서버와 클라이언트 양쪽에서 사용하는 타입이므로 apis/client/document.ts가 아닌
type/Document.type.ts에 선언합니다. 제목 목록 BFF 응답 타입도 실제 응답에 맞춰
DocumentTitle[]로 명확히 했습니다.
