+++ client/src/app/api/get-document-title-list/route.ts
@@ -1,14 +1,14 @@
 'use server';
 
-import {TitleAndUUID} from '@apis/client/document';
 import {getDocumentTitlesServer} from '@apis/server/document';
+import {DocumentTitle} from '@type/Document.type';
 import {ApiResponseType} from '@type/http.type';
 import {NextResponse} from 'next/server';
 
 export const GET = async () => {
   const documents = await getDocumentTitlesServer();
 
-  const response: ApiResponseType<TitleAndUUID[]> = {
+  const response: ApiResponseType<DocumentTitle[]> = {
     data: documents,
     code: 'SUCCESS',
   };
